requirements.txt
python-telegram-bot>=21.0
anthropic>=0.39.0
openai>=1.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
aiofiles>=23.0.0
